v0.71.2 - Update deps and docker image

+++ packages/scripts/src/helpers/docs/typedoc.ts
@@ -1,12 +1,14 @@
 import path from 'path';
 import fse from 'fs-extra';
-import { Application } from 'typedoc';
-import { JsxEmit, ModuleKind, ModuleResolutionKind, ScriptTarget } from 'typescript';
+import { Application, TSConfigReader } from 'typedoc';
+import {
+    JsxEmit, ModuleKind, ModuleResolutionKind, ScriptTarget
+} from 'typescript';
 import { PackageInfo } from '../interfaces';
 import { listMdFiles, getName, writeIfChanged } from '../misc';
 import signale from '../signale';
 
-const app = new Application();
+
 
 function isOverview(filePath: string): boolean {
     return path.basename(filePath, '.md') === 'overview';
@@ -23,7 +25,7 @@ async function writeDocFile(filePath: string, { title, sidebarLabel }: { title: 
 
     // fix paths
     contents = contents
-        .replace(/(\]\([\w\.\/]*)index\.md/g, '$1overview.md');
+        .replace(/(\]\([\w\.\/]*)README\.md/g, '$1overview.md');
     // build final content
     contents = `---
 title: ${title}
@@ -51,10 +53,10 @@ function getAPIName(overview: string, outputDir: string, filePath: string) {
 }
 
 async function fixDocs(outputDir: string, { displayName }: PackageInfo) {
-    const overviewFilePath = listMdFiles(outputDir).find((filePath) => path.basename(filePath, '.md') === 'index');
+    const overviewFilePath = listMdFiles(outputDir).find((filePath) => path.basename(filePath, '.md') === 'README');
     if (!overviewFilePath) {
         signale.error(
-            'Error: Package documentation was not generated correctly',
+            'Error: Package documentation was not generated correctly' +
             ", this means the package my not work with the typedoc's version of TypeScript."
         );
         return;
@@ -82,48 +84,54 @@ async function fixDocs(outputDir: string, { displayName }: PackageInfo) {
     await Promise.all(promises);
 }
 
-export async function generateTSDocs(pkgInfo: PackageInfo, outputDir: string) {
+
+
+export async function generateTSDocs(pkgInfo: PackageInfo, outputDir: string): Promise<void> {
     signale.await(`building typedocs for package ${pkgInfo.name}`);
 
     const cwd = process.cwd();
     try {
         process.chdir(pkgInfo.dir);
+        const app = new Application();
+        app.options.addReader(new TSConfigReader());
         app.bootstrap({
             name: pkgInfo.name,
-            target: ScriptTarget.ESNext,
             tsconfig: path.join(pkgInfo.dir, 'tsconfig.json'),
-            // @ts-expect-error
-            platform: 'docusaurus' as any,
-            mode: 'file',
             theme: 'markdown',
             exclude: ['test', 'node_modules'],
             excludePrivate: true,
             excludeExternals: true,
-            excludeNotExported: true,
-            experimentalDecorators: true,
-            skipLibCheck: true,
-            esModuleInterop: true,
-            ignoreCompilerErrors: true,
-            strict: false,
-            jsx: JsxEmit.React,
-            moduleResolution: ModuleResolutionKind.NodeJs,
-            module: ModuleKind.CommonJS,
             hideGenerator: true,
             readme: 'none',
-        })
-        const inputFiles = [...app.expandInputFiles(['src'])];
+        });
+        app.options.setValue("entryPoints", app.expandInputFiles(['src']));
+
+        if (app.logger.hasErrors()) {
+            signale.error(`found errors typedocs for package ${pkgInfo.name}`);
+            return;
+        }
+        const project = app.convert();
+        if (!project) {
+            signale.error(`invalid typedocs for package ${pkgInfo.name}`);
+            return;
+        }
 
         if (fse.existsSync(outputDir)) {
             await fse.emptyDir(outputDir);
         }
         await fse.ensureDir(outputDir);
-        // Project may not have converted correctly
-        // Rendered docs
-        app.generateDocs(inputFiles, outputDir);
+
+        await app.generateDocs(project, outputDir);
+
+        if (app.logger.hasErrors()) {
+            signale.error(`found errors when generating typedocs for package ${pkgInfo.name}`);
+            return;
+        }
 
         await fixDocs(outputDir, pkgInfo);
-    } finally {
+
         signale.success(`generated docs for package ${pkgInfo.name}`);
+    } finally {
         process.chdir(cwd);
     }
 }
